refactor(core): split agentTools.ts into agentTools/ modules
agentTools.ts (619 lines) exceeded the 500-line limit. Split into
cohesive modules under agentTools/: execTool (bash exec), fsMounts
(mount resolution + glob matching), fileTools (read_file/list_files),
grepTool, skillTool (read_skill), and researchTools (buildResearchTools
aggregator). agentTools.ts is now a pure re-export barrel preserving
the original public API and tool registration order, so no consumer
(including apps/pro) needs to change.
